Item Database, Loot Regions, Inventory, and Item scripts are ported.
Inventory view is in progress

+++ Data/Scripts/ItemSystems_Database.lua
@@ -4,8 +4,11 @@
 
     Reads and indexes the raw data scripts. Provides factory methods for creating items.
 ]]
+
+
 local Item = require(script:GetCustomProperty("Item"))
-local LOOT_TABLE = script:GetCustomProperty("LootTable"):WaitForObject()
+local LOOT_TABLE = script:GetCustomProperty("LootTable"):WaitForObject():GetChildren()
+local SALVAGE_TABLE = script:GetCustomProperty("SalvageTable"):WaitForObject():GetChildren()
 local REGISTERED_ITEMS = script:GetCustomProperty("RegisteredItems"):WaitForObject()
 
 
@@ -13,13 +16,33 @@ local REGISTERED_ITEMS = script:GetCustomProperty("RegisteredItems"):WaitForObje
 local LOAD_FRAME_LIMIT = 10
 
 -- If true, when the game loads it will log all the catalogs and their items that are registered to that catalog.
-local DEBUGLOGLOAD = true 
+local DEBUGLOGLOAD = false 
 
 local DATA_ITEMS = {}
 
-for _, item in pairs(REGISTERED_ITEMS:GetCustomProperties()) do
-    table.insert(DATA_ITEMS,item)
+local function HasRequiredProperties(item)
+    if item:GetCustomProperty("Name") and
+        item:GetCustomProperty("Icon") and
+        item:GetCustomProperty("ItemType") and
+        item:GetCustomProperty("Rarity") then
+        return true
+    else
+        return false
+    end
 end
+
+-- Recursively load the item database.
+local function LoadItems_R(root)
+    if not root then return end
+    for _, item in pairs(root:GetChildren()) do
+        if HasRequiredProperties(item) then
+            table.insert(DATA_ITEMS,item)
+        elseif #item:GetChildren() > 0 then
+            LoadItems_R(item)
+        end
+    end
+end
+LoadItems_R(REGISTERED_ITEMS)
 
 ---------------------------------------------------------------------------------------------------------
 -- PUBLIC
@@ -32,40 +55,131 @@ function Database:WaitUntilLoaded()
     end
 end
 
-function Database:CreateItemDropFromName(itemName)
+-------------- Item Factory Methods --------------
+-- These methods are for creating items without stats.
+-- This is useful for adding items to the inventory or checking for items.
+
+function Database:GetItemFromName(itemName)
     local itemData = Database:FindItemDataByName(itemName)
     local item = Item.New(itemData)
+    if not itemData then
+        warn(string.format("Failed to create item - %s from database. Are you sure this is the right name?",itemName))
+    end
+    return item
+end
+
+function Database:GetItemFromMUID(muid)
+    local itemData = Database:FindItemDataByFullMUID(muid) or Database:FindItemDataByMUID(muid)
+    if not itemData then
+        warn(string.format("Failed to create item - %s from database. Are you sure this is the right muid?",muid))
+    end
+    local item = Item.New(itemData)
+    return item
+end
+
+function Database:GetItemFromIndex(index)
+    local itemData = Database:FindItemDataByIndex(index)
+    local item = Item.new(itemData)
+    if not itemData then
+        warn(string.format("Failed to create item - %s from database. Are you sure this is the right index?",index))
+    end
+    return item
+end
+
+function Database:CreateItemFromHash(itemHash)
+    local item = Item.FromHash(self, itemHash)
+    if not item then
+        warn(string.format("Failed to create item - %s from database. This may not be a valid hash.",itemHash))
+    end
+    return item
+end
+
+-------------- Loot Factory Methods ---------------
+-- These methods are for creating items that roll for stats
+-- This is useful for creating items that roll for stats.
+-- This framework does not use stats on items.
+
+function Database:CreateLootItemFromName(itemName)
+    local itemData, min, max = Database:FindItemDataByName(itemName)
+    local item = Item.New(itemData)
+    if not itemData then
+        warn(string.format("Failed to create item - %s from database. Are you sure this is the right name?",itemName))
+    end
+    if min and max then
+        item:SetStackSize(math.random(min,max))
+    end
     self:_RollItemStats(item)
     return item
 end
 
-function Database:CreateItemFromDrop(dropKey)
-    local itemData = self:_RollDrop(dropKey)
-    assert(itemData, string.format("Could not drop item from %s group as the group may not exist in LootTables group.", dropKey))
+function Database:CreateLootItemFromMUID(muid)
+    local itemData, min, max = Database:FindItemDataByFullMUID(muid) or Database:FindItemDataByMUID(muid)
     local item = Item.New(itemData)
+    if not itemData then
+        warn(string.format("Failed to create item - %s from database. Are you sure this is the right muid?",muid))
+    end
+    if min and max then
+        item:SetStackSize(math.random(min,max))
+    end
+    self:_RollItemStats(item)
+    return item
+end
+
+function Database:CreateLootItemFromIndex(index)
+    local itemData, min, max = Database:FindItemDataByIndex(index)
+    local item = Item.new(itemData)
+    if not itemData then
+        warn(string.format("Failed to create item - %s from database. Are you sure this is the right index?",index))
+    end
+    if min and max then
+        item:SetStackSize(math.random(min,max))
+    end
+    self:_RollItemStats(item)
+    return item
+end
+
+function Database:CreateLootItemFromDropKey(dropKey)
+    local itemData, min, max = self:_RollDrop(dropKey)
+    if not itemData then
+        warn(string.format([[ 
+        Failed to create item from - %s loot table. Does this loot table exist in item registry?
+        ]],dropKey))
+        return nil
+    end
+    if itemData == nil or itemData == "Empty" then return end
+    --assert(itemData, string.format("Could not drop an item from %s loot table as it does not exist.", dropKey))
+    local item = Item.New(itemData)
+    if min and max then
+        item:SetStackSize(math.random(min,max))
+    end
     self:_RollItemStats(item)
     return item 
 end
 
-function Database:CreateItemFromHash(itemHash)
-    return Item.FromHash(self, itemHash)
-end
-
-function Database:CreateItemFromItemData(itemData)
+function Database:CreateLootItemFromItemData(itemData)
     itemData._RollStats = self:_GetRollFunction(itemData.statKey)
     local item = Item.New(itemData)
     self:_RollItemStats(item)
     return item 
 end
+-----------------------------------------------
 
+-- Returns a table containing the data of an item when provided with the name of the item. Example: 1
 function Database:FindItemDataByIndex(itemIndex)
     return self.itemDatasByIndex[itemIndex]
 end
 
+-- Returns a table containing the data of an item when provided with the name of the item. Example: Wood
 function Database:FindItemDataByName(itemName)
     return self.itemDatasByName[itemName]
 end
 
+-- Returns a table containing the data of an item when provided with a full MUID. Example: F2F89CB6DC0893B7:ITEM_Misc_Wood
+function Database:FindItemDataByFullMUID(itemMUID)
+    return self.itemDatasByMUIDFull[itemMUID]
+end
+
+-- Returns a table containing the data of an item when provided with a MUID. Example: F2F89CB6DC0893B7
 function Database:FindItemDataByMUID(itemMUID)
     return self.itemDatasByMUID[itemMUID]
 end
@@ -82,7 +196,6 @@ function Database:_Init()
     Task.Spawn(function()
         self:_LoadItems()
         self:_LoadDrops()
-        self:_LoadAssetDerivedInformation()
         self.isLoaded = true
     end)
 end
@@ -118,30 +231,49 @@ function Database:_LoadStats(item)
     end
 end
 
+-- Finds the assiociated salvage components folder for the item and loads the componenents into a table
+function Database:_LoadSalvageComponents(itemTemplate)
+    for _, salvagableItem in pairs(SALVAGE_TABLE) do
+        if salvagableItem:GetCustomProperty("ItemToSalvage") == itemTemplate then
+            local resources = {}
+            for i, component in pairs(salvagableItem:GetChildren()) do
+                local componentProperties = component:GetCustomProperties()
+                assert(componentProperties["ItemProduct"],string.format("%s in salvagables folder contains a component that does not have an assigned item.",salvagableItem.name))
+                local item = componentProperties["ItemProduct"] -- The item asset reference
+                local amount = componentProperties["Amount"] or 1 -- The amount of that item
+                table.insert(resources,{ item = item, amount = amount })
+            end
+            return resources ~= {} and resources or nil
+        end
+    end
+end
+
 function Database:_LoadItems()
     self.itemDatasByIndex = {}
     self.itemDatasByName = {}
+    self.itemDatasByMUIDFull = {}
     self.itemDatasByMUID = {}
     local index = 1
     if DEBUGLOGLOAD then print("Loading Items ----------------------------------------------------- \n") end -- Debug
-    for _, itemTemplate in ipairs(DATA_ITEMS) do
-        local tempItem = World.SpawnAsset(itemTemplate)
-        local muid = tempItem.sourceTemplateId
-        local name = tempItem.name
-        local propName = tempItem:GetCustomProperty("Name")
-        local propIcon = tempItem:GetCustomProperty("Icon")
-        local propMaxStackableSize = tempItem:GetCustomProperty("MaxStackableSize")
-        local propItemType = tempItem:GetCustomProperty("ItemType")
-        local propDescription = tempItem:GetCustomProperty("Description")
-        local propRarity = tempItem:GetCustomProperty("Rarity")
-        local propLevelRequirement = tempItem:GetCustomProperty("LevelRequirement")
-        local propEquipmentStance = tempItem:GetCustomProperty("EquipmentStance")
-        local propConsumptionEffect = tempItem:GetCustomProperty("ConsumptionEffect")
-        self:_LoadStats(tempItem)
-        tempItem:Destroy()
+    for index, item in ipairs(DATA_ITEMS) do
+        if index % LOAD_FRAME_LIMIT == 0 then Task.Wait() end
+        local itemMUID = item:GetCustomProperty("Item")
+        local propName = item:GetCustomProperty("Name")
+        local propIcon = item:GetCustomProperty("Icon")
+        local propMaxStackableSize = item:GetCustomProperty("MaxStackableSize")
+        local propItemType = item:GetCustomProperty("ItemType")
+        local propDescription = item:GetCustomProperty("Description")
+        local propRarity = item:GetCustomProperty("Rarity")
+        local propLevelRequirement = item:GetCustomProperty("LevelRequirement")
+        local propEquipmentStance = item:GetCustomProperty("EquipmentStance")
+        local propConsumptionEffect = item:GetCustomProperty("ConsumptionEffect")
+        local propBackpackSlotCount = item:GetCustomProperty("BackpackSlotCount")
+        local isBackpack = propBackpackSlotCount and true or false
+        local salvageComponents = self:_LoadSalvageComponents(itemMUID) -- Load salvage components for the item
+        self:_LoadStats(item) -- Load stats from the item
 
-        if DEBUGLOGLOAD then print(name) end -- Debug
-        if DEBUGLOGLOAD then print("|",muid, "    =",propName) end -- Debug
+        if DEBUGLOGLOAD then print(propName) end -- Debug
+        if DEBUGLOGLOAD then print("|",itemMUID, "    =",propName) end -- Debug
 
         if propMaxStackableSize then
             assert(tonumber(propMaxStackableSize) <= 2^12, string.format("item stack size is too large - %s", propName))
@@ -152,7 +284,7 @@ function Database:_LoadItems()
         end
 
         assert(not self.itemDatasByName[propName], string.format("duplicate item name is not allowed - %s check your registered items for duplicates", propName))
-        assert(not self.itemDatasByMUID[muid], string.format("duplicate item MUID is not allowed - %s", muid))
+        assert(not self.itemDatasByMUIDFull[muid], string.format("duplicate item MUID is not allowed - %s", muid))
         assert(Item.SLOT_CONSTRAINTS[propItemType], string.format("unrecognized item type - %s from %s add your item type to the ItemTypes folder in ItemRegistry", propItemType, propName))
         assert(Item.RARITIES[propRarity], string.format("unrecognized item rarity - %s check ItemSystems_ItemThemes to make sure this rarity exist.", propRarity))
 
@@ -168,16 +300,21 @@ function Database:_LoadItems()
             levelRequirement = propLevelRequirement,
             stance = stance,
             isEquippable = isEquippable,
+            isBackpack = isBackpack,
+            backpackSlotCount = propBackpackSlotCount,
             maxStackSize = maxStackSize,
-            muid = muid,
+            muid = itemMUID:match("^(.+):"),
             description = propDescription or "",
             consumptionEffect = propConsumptionEffect,
-            _RollStats = Database:_GetRollFunction(name)
+            salvageComponents = salvageComponents,
+            _RollStats = Database:_GetRollFunction(propName)
         }
+
         index = index + 1
         self.itemDatasByIndex[itemData.index] = itemData
         self.itemDatasByName[itemData.name] = itemData
-        self.itemDatasByMUID[itemData.muid] = itemData
+        self.itemDatasByMUIDFull[itemMUID] = itemData
+        self.itemDatasByMUID[itemMUID:match("^(.+):")] = itemData
         if DEBUGLOGLOAD then print("\n") end -- Debug
     end
     if DEBUGLOGLOAD then print("Loading Items Ended -----------------------------------------------\n") end -- Debug
@@ -223,59 +360,38 @@ end
 function Database:_LoadDrops()
     self.itemDropTables = {}
     self.itemDropKeys = {}
-    for _, lootTable in pairs(LOOT_TABLE:GetChildren()) do
+    for _, lootTable in pairs(LOOT_TABLE) do
         if not self.itemDropTables[lootTable.name] then
             self.itemDropTables[lootTable.name] = { cumulativeLikelihood = 0 }
             table.insert(self.itemDropKeys, lootTable.name)
         end
         for _, lootDrop in pairs(lootTable:GetChildren()) do
-            local item = lootDrop:GetCustomProperty("Item")
+            local propItem = lootDrop:GetCustomProperty("Item")
+            local item = propItem and propItem:GetObject() or "Empty"
             local likelihood = lootDrop:GetCustomProperty("Likelihood")
-            local tempObject = World.SpawnAsset(item)
-            local itemName = tempObject:GetCustomProperty("Name")
-            assert(self:FindItemDataByName(itemName), string.format("loot drop references unknown item - %s at loot table - %s make sure your item is registered.", itemName, lootTable.name))
-            assert(likelihood, string.format("loot drop missing likelihood custom property - %s at loot table - %s", itemName, lootTable.name))
-            local dropTable = self.itemDropTables[lootTable.name]
-            local dropInfo = { itemName = itemName, likelihood = tonumber(likelihood) }
-            table.insert(dropTable, dropInfo)
-            dropTable.cumulativeLikelihood = dropTable.cumulativeLikelihood + dropInfo.likelihood
-            tempObject:Destroy()
+            local minQuantity = lootDrop:GetCustomProperty("MinimumStackWhenDropped")
+            local maxQuantity = lootDrop:GetCustomProperty("MaximumStackWhenDropped")
+            if item == "Empty" and likelihood then
+                local dropTable = self.itemDropTables[lootTable.name]
+                local dropInfo = { itemName = "Empty", likelihood = tonumber(likelihood)}
+                table.insert(dropTable, dropInfo)
+                dropTable.cumulativeLikelihood = dropTable.cumulativeLikelihood + dropInfo.likelihood
+            elseif item ~= "Empty" then
+                local itemName = item:GetCustomProperty("Name")
+                if not self:FindItemDataByName(itemName) then
+                    warn(string.format("%s group inside the loot table - %s. The group's item property does not reference a registered item.",lootDrop.name,lootTable.name))
+                end
+                if not likelihood then
+                    warn(string.format("Likelihood property is missing from %s in loot table %s",lootDrop.name,lootTable.name))
+                end
+                if itemName and likelihood then
+                    local dropTable = self.itemDropTables[lootTable.name]
+                    local dropInfo = { itemName = itemName, likelihood = tonumber(likelihood), min = minQuantity, max = maxQuantity }
+                    table.insert(dropTable, dropInfo)
+                    dropTable.cumulativeLikelihood = dropTable.cumulativeLikelihood + dropInfo.likelihood
+                end
+            end
         end
-    end
-
-    -- for dropKey,lootTable in pairs() do
-
-    --     if not self.itemDropTables[dropKey] then
-    --         self.itemDropTables[dropKey] = { cumulativeLikelihood = 0 }
-    --         table.insert(self.itemDropKeys, dropKey)
-    --     end
-    --     for i, row in pairs(lootTable) do
-    --         assert(row.ItemName, string.format("loot drop missing name at row - %d at loot table - %s", i, dropKey))
-    --         assert(self:FindItemDataByName(row.ItemName), string.format("loot drop references unknown item - %s at loot table - %s", row.ItemName, dropKey))
-    --         assert(row.Likelihood, string.format("loot drop missing likelihood - %s at loot table - %s", row.ItemName, dropKey))
-
-    --         local dropTable = self.itemDropTables[dropKey]
-    --         local dropInfo = { itemName = row.ItemName, likelihood = tonumber(row.Likelihood) }
-    --         table.insert(dropTable, dropInfo)
-    --         dropTable.cumulativeLikelihood = dropTable.cumulativeLikelihood + dropInfo.likelihood
-    --     end
-    -- end
-end
-
-function Database:_LoadAssetDerivedInformation()
-    local itemCount = #self.itemDatasByIndex
-    local itemsPerFrame = math.ceil(itemCount / LOAD_FRAME_LIMIT)
-    for index,itemData in ipairs(self.itemDatasByIndex) do
-        if index % itemsPerFrame == 0 then
-            Task.Wait()
-        end
-        local tempObject = World.SpawnAsset(itemData.muid)
-        itemData.icon = tempObject:GetCustomProperty("Icon")
-        assert(itemData.icon, string.format("item template %s missing icon property", itemData.muid))
-        itemData.iconRotation = tempObject:GetCustomProperty("IconRotation")
-        itemData.iconColorTint = tempObject:GetCustomProperty("IconColorTint")
-        itemData.animationStance = tempObject:GetCustomProperty("AnimationStance")
-        tempObject:Destroy()
     end
 end
 
@@ -285,7 +401,11 @@ function Database:_RollDrop(dropKey)
         local roll = math.random() * dropTable.cumulativeLikelihood
         for _,dropInfo in ipairs(dropTable) do
             if roll <= dropInfo.likelihood then
-                return self:FindItemDataByName(dropInfo.itemName)   
+                local item, min, max = self:FindItemDataByName(dropInfo.itemName), dropInfo.min, dropInfo.max 
+                if dropInfo.itemName == "Empty" then
+                    return "Empty", min, max
+                end
+                return item, min, max
             end
             roll = roll - dropInfo.likelihood
         end
